requirements.txt
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
scipy==1.11.1
jupyterlab==4.0.2
jupyter==1.0.0
pyyaml==6.0